% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_age.R
\name{aggregate_by_age}
\alias{aggregate_by_age}
\title{Aggregate linelist data into counts by age group}
\usage{
aggregate_by_age(
  data,
  age_group = "age_group",
  split_by = "sex",
  stack_by = NULL,
  proportional = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{a data frame}

\item{age_group}{the name of a column in the data frame that defines the age
group categories. Defaults to "age_group"}

\item{split_by}{the name of a column in the data frame that defines the
the bivariate column. Defaults to "sex". See NOTE}

\item{stack_by}{the name of the column in the data frame to use for shading
the bars}

\item{proportional}{If \code{TRUE}, bars will represent proportions of cases out
of the entire population. Otherwise (\code{FALSE}, default), bars represent
case counts}

\item{na.rm}{If \code{TRUE}, this removes NA counts from the age groups. Defaults
to \code{TRUE}.}
}
\value{
a data frame with counts aggregated by age
}
\description{
Aggregate linelist data into counts by age group
}
\examples{
set.seed(2018 - 01 - 15)
ages <- cut(sample(80, 150, replace = TRUE),
  breaks = c(0, 5, 10, 30, 90), right = FALSE
)
sex <- sample(c("Female", "Male"), 150, replace = TRUE)
gender <- sex
gender[sample(5)] <- "NB"
ill <- sample(c("case", "non-case"), 150, replace = TRUE)
dat <- data.frame(
  AGE = ages,
  sex = factor(sex, c("Male", "Female")),
  gender = factor(gender, c("Male", "NB", "Female")),
  ill = ill,
  stringsAsFactors = FALSE
)
aggregate_by_age(dat, 
                 age_group = AGE,
                 split_by = gender, 
                 stack_by = ill,
                 proportional = TRUE,
                 na.rm = FALSE)
}
