% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_2018}
\alias{us_2018}
\alias{us_2008}
\alias{us_ins_2018}
\alias{us_ins_2008}
\alias{us_gen_2018}
\alias{us_gen_2008}
\title{US Census data for population, age, and gender}
\format{a data frame with columns and rows}
\source{
\url{https://census.gov/data/tables/2018/demo/age-and-sex/2018-age-sex-composition.html}
\url{https://census.gov/data/tables/2008/demo/age-and-sex/2008-age-sex-composition.html}
}
\usage{
us_2018

us_2008

us_ins_2018

us_ins_2008

us_gen_2018

us_gen_2008
}
\description{
US Census data for population, age, and gender
}
\keyword{datasets}
